Sensor pins assignable; LDR options; bugfix
(ESP8266 only) The pins for sensors can be chosen in the options and will be stored in the configuration file.

LDR type and pulldown resistor value can be set in the options.

MP3Player sometimes cannot handle two commands in a row ("volume", then "play"). Insert delay if there's a second command.