Rearrange and fix WebUI
Rearranged some of the settings on Options & Sensor pages.

On the dashboard, Button Data is not provided in the corresponding section. Might be a typo which I corrected. @o0shojo0o Please check if this area gets messages now (I integrated a first message to be sent on ws connect).