Rearrange and fix WebUI (#115)
* Rearrange and fix WebUI

Rearranged some of the settings on Options & Sensor pages.

On the dashboard, Button Data is not provided in the corresponding section. Might be a typo which I corrected. @o0shojo0o Please check if this area gets messages now (I integrated a first message to be sent on ws connect).

* Add button actions for MP3 control

Buttons can be assigned Next, Previous and Play/Pause actions for the MP3 player.

Might still need some fixing for the problems with some DFPlayer modules.

* Offer Zigzag animation and random animation for Clock Auto Fallback

Offer new animation types for Auto Fallback as well. (BitmapAnimation is not available, as it would require a defined bitmap.) Colors of ZigZag Animation will be clock colors.