Increase interval between reconnect attempts; correct LWT (#135)
In my experience, a connection attempt to a non-existant broker needs about 5 seconds before timing out. With a reconnect invertal of 5 seconds, the next connection attempt will follow immediately.
As the attempt is blocking, API will not respond during the attempts. In other words: you can't disable MQTT if the broker is failing, because there is no chance to open the WebUI.

Corrected a typo in the Last Will message and corrected the topic for the Last Will message (shouldn't be a global state).